+++ src/clients/python/sams_client/endpoints/images.py
@@ -9,6 +9,65 @@
 # AUTHORS and LICENSE files distributed with this source code, or
 # at https://www.sourcefabric.org/superdesk/license
 
+"""
+
+A set a collection of *Assets*. It is used as a way to organise and restrict access to
+the *Assets* it holds.
+
+All access to the binary or metadata of assets will go through this service. There must
+be at least one Set configured in the system.
+
+A single Set can be designated as the default Set. When access is requested to an asset
+without a set being provided, the default set will be used.
+
+Usage
+-----
+The Set service instance can be found under :data:`sams.sets.service` and used all::
+
+    from sams_client import SamsClient, IAsset, IAssetTag
+
+    client = SamsClient()
+
+    # Load the image file
+    with open("example_image.png", "rb") as image_file:
+        asset_binary = image_file.read()
+
+    set_id = client.sets.search().json()["_items"][0]["_id"]
+
+    # Upload a new Image Asset
+    response = client.assets.create(
+        docs=IAsset(
+            set_id=set_id,
+            filename="example_image.png",
+            name="test image renditions",
+            description="example image to test renditions",
+            tags=[
+                IAssetTag(code="test", name="test"),
+                IAssetTag(code="rendition", name="rendition")
+            ]
+        ),
+        files={"binary": asset_binary},
+        external_user_id="test_user_id"
+    )
+    new_asset = response.json()
+
+    # Generate a rendition
+    client.images.generate_rendition(
+        new_asset["_id"],
+        width=640,
+        height=640,
+        keep_proportions=True
+    )
+
+    # Download the newly created rendition
+    image_rendition = client.images.download(
+        new_asset["_id"],
+        width=640,
+        height=640,
+        keep_proportions=True
+    )
+"""
+
 from typing import Dict, Any, Callable, Union, Optional
 
 from bson import ObjectId
@@ -36,7 +95,19 @@ class SamsImagesEndpoint:
         keep_proportions: Optional[bool] = True,
         headers: Dict[str, Any] = None,
         callback: Callable[[requests.Response], requests.Response] = None
-    ):
+    ) -> requests.Response:
+        r"""Download an Image, optionally providing image dimensions
+
+        :param str item_id: The Asset ID
+        :param int width: Desired image width (optional)
+        :param int height: Desired image height (optional)
+        :param bool keep_proportions: If `true`, keeps image width/height ratio
+        :param dict headers: Dictionary of headers to apply
+        :param callback: A callback function to manipulate the response
+        :rtype: requests.Response
+        :return: The Asset binary, optionally resized
+        """
+
         params = {}
 
         if width:
@@ -56,14 +127,26 @@ class SamsImagesEndpoint:
         )
 
     def generate_rendition(
-            self,
-            item_id: Union[ObjectId, str],
-            width: Optional[int] = None,
-            height: Optional[int] = None,
-            keep_proportions: Optional[bool] = True,
-            headers: Dict[str, Any] = None,
-            callback: Callable[[requests.Response], requests.Response] = None
-    ):
+        self,
+        item_id: Union[ObjectId, str],
+        width: Optional[int] = None,
+        height: Optional[int] = None,
+        keep_proportions: Optional[bool] = True,
+        headers: Dict[str, Any] = None,
+        callback: Callable[[requests.Response], requests.Response] = None
+    ) -> requests.Response:
+        r"""Generates an Image rendition
+
+        :param str item_id: The Asset ID
+        :param int width: Desired image width (optional)
+        :param int height: Desired image height (optional)
+        :param bool keep_proportions: If `true`, keeps image width/height ratio
+        :param dict headers: Dictionary of headers to apply
+        :param callback: A callback function to manipulate the response
+        :rtype: requests.Response
+        :return: 200 status code if rendition generated successfully
+        """
+
         params = {}
 
         if width: